Finalized controller handler.
Also implemented render system drawing loop.

+++ ArgosProject/src/stdafx.h
@@ -136,7 +136,7 @@ namespace app
 		using KeyHandler = Keyhandler<KeyCode>;
 		using MouseButtonCode = std::uint8_t;
 		using MouseHandler = Mousehandler<MouseButtonCode>;
-		using ControllerIndex = std::int16_t;
+		using ControllerIndex = std::int32_t;
 		using ControllerButtonCode = SDL_GameControllerButton;
 		using ControllerAxisCode = SDL_GameControllerAxis;
 		using ControllerAxisValue = Sint16;
